+++ core/common/regexp.cpp
@@ -24,7 +24,7 @@ std::vector<std::string> Regexp::exec(const std::string& str) const
     {
         std::vector<std::string> v;
 
-        for (int i = 0; i < m.size(); i++)
+        for (size_t i = 0; i < m.size(); i++)
             v.push_back(m[i].str());
         return v;
     }else
